# openapi/railwaysMySQL-api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8081
  description: Generated server url
paths:
...
  /api/trains/{trainId}:
    get:
      tags:
      - trains
      summary: Get an existing train
      operationId: getSingleTrain
      parameters:
      - name: trainId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: No train to get
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Train'
    put:
      tags:
      - trains
      summary: Update an existing train
      operationId: updateSingleTrain
      parameters:
      - name: trainId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Train'
        required: true
      responses:
        "200":
          description: Train successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Train'
        "404":
          description: No train to update
    delete:
      tags:
      - trains
      summary: Delete an existing train
      operationId: deleteTrain
      parameters:
      - name: trainId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: No train to delete
        "204":
          description: No content
...
  /api/trains:
    get:
      tags:
      - trains
      summary: Get trains list with optional filters
      operationId: findTrainPage
      parameters:
      - name: companyName
        in: query
        required: false
        schema:
          type: string
      - name: trainType
        in: query
        required: false
        schema:
          type: string
          enum:
          - REGIONALE
          - INTERCITY
          - FRECCIAROSSA
      - name: offset
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: limit
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
    post:
      tags:
      - trains
      summary: Add a new train
      operationId: createTrain
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Train'
        required: true
      responses:
        "409":
          description: Already exists a train with id in body
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Train'
  /api/trains/{trainId}/wagon/{wagonId}:
    post:
      tags:
      - wagons-trains
      summary: Add or rewrite a new association between wagon and train
      operationId: createOrUpdateTrainWagonAssociation
      parameters:
      - name: trainId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: wagonId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "409":
          description: Already exists an association between ids in path
        "201":
          description: Created
  /api/trains/{trainId}/routes/{routeId}:
    post:
      tags:
      - routes-trains
      summary: Add a new association between route and train
      operationId: createTrainRouteAssociation
      parameters:
      - name: trainId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: routeId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "409":
          description: Already exists an association between ids in path
        "201":
          description: Created
    delete:
      tags:
      - routes-trains
      summary: Delete an existing association between train and route
      operationId: deleteTrainRouteAssociation
      parameters:
      - name: trainId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: routeId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: No association to delete
        "204":
          description: No content
...
  /api/trains/{trainId}/wagons:
    get:
      tags:
      - wagons-trains
      summary: Get wagons list of a specific train with optional filters
      operationId: findWagonsOfTrainPage
      parameters:
      - name: trainId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: wagonType
        in: query
        required: false
        schema:
          type: string
          enum:
          - MOTOR
          - PASSENGER
          - BED
          - RESTAURANT
      - name: kitchenType
        in: query
        required: false
        schema:
          type: string
      - name: fuelType
        in: query
        required: false
        schema:
          type: string
          enum:
          - ELECTRIC
          - DIESEL
      - name: bathroom
        in: query
        required: false
        schema:
          type: boolean
      - name: wagonClass
        in: query
        required: false
        schema:
          type: string
          enum:
          - PRIMA
          - SECONDA
          - ECONOMY
      - name: bedType
        in: query
        required: false
        schema:
          type: string
      - name: minimumTables
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: minimumBeds
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: minimumSeats
        in: query
        required: false
        schema:
          type: integer
          format: int32
      - name: offset
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: limit
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        "404":
          description: No train to get
  /api/trains/{trainId}/routes:
    get:
      tags:
      - routes-trains
      summary: Get routes list of a specific train with optional filters
      operationId: findRoutesOfTrainPage
      parameters:
      - name: trainId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: startStation
        in: query
        required: false
        schema:
          type: string
      - name: destinationStation
        in: query
        required: false
        schema:
          type: string
      - name: binaryType
        in: query
        required: false
        schema:
          type: string
      - name: minimumLength
        in: query
        required: false
        schema:
          type: number
          format: double
      - name: maxLength
        in: query
        required: false
        schema:
          type: number
          format: double
      - name: offset
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: limit
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page'
        "404":
          description: No train to get
components:
  schemas:
...
    Train:
      type: object
      properties:
        id:
          type: integer
          format: int32
        type:
          type: string
          enum:
          - REGIONALE
          - INTERCITY
          - FRECCIAROSSA
        company:
          type: string
        routes:
          uniqueItems: true
          type: array
          items:
            properties:
              routeIDs:
                type: number
                format: int32
        wagons:
          uniqueItems: true
          type: array
          items:
            properties:
              wagonIDs:
                type: number
                format: int32
    Page:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            type: object
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/Sort'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/Pageable'
        empty:
          type: boolean
    Pageable:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/Sort'
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
        unpaged:
          type: boolean
        paged:
          type: boolean
    Sort:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean